move config type to client
