+++ plugin/evm/client/client.go
@@ -40,7 +40,7 @@ type Client interface {
 	MemoryProfile(ctx context.Context, options ...rpc.Option) error
 	LockProfile(ctx context.Context, options ...rpc.Option) error
 	SetLogLevel(ctx context.Context, level slog.Level, options ...rpc.Option) error
-	// GetVMConfig(ctx context.Context, options ...rpc.Option) (*Config, error)
+	GetVMConfig(ctx context.Context, options ...rpc.Option) (*Config, error)
 }
 
 // Client implementation for interacting with EVM [chain]
@@ -299,9 +299,13 @@ func (c *client) SetLogLevel(ctx context.Context, level slog.Level, options ...r
 	}, &api.EmptyReply{}, options...)
 }
 
+type ConfigReply struct {
+	Config *Config `json:"config"`
+}
+
 // GetVMConfig returns the current config of the VM
-// func (c *client) GetVMConfig(ctx context.Context, options ...rpc.Option) (*Config, error) {
-// 	res := &ConfigReply{}
-// 	err := c.adminRequester.SendRequest(ctx, "admin.getVMConfig", struct{}{}, res, options...)
-// 	return res.Config, err
-// }
+func (c *client) GetVMConfig(ctx context.Context, options ...rpc.Option) (*Config, error) {
+	res := &ConfigReply{}
+	err := c.adminRequester.SendRequest(ctx, "admin.getVMConfig", struct{}{}, res, options...)
+	return res.Config, err
+}
